Linear Search uploaded
Linear search is a simple method for finding a specific value within a list or array. In a linear search, each element of the list is checked one by one, starting from the beginning, until the desired element is found or the end of the list is reached. It is a basic and straightforward approach. It is also known as sequential search.